Fix extra slash in `Machine.spec.providerID` (#26)
During the reverse engineering of the GCP and AWS providers I erroneously assumed the provider IDs needs three slashes behind `cloudscale:`. In reality it is `gcp://`+`/instance` for those providers. Cloudscale does not have this extra slash.

All machines should be updated to the correct ID on the first reconcile with the fixed code. Node objects get the correct ID from the cloudscale CCM already.